// src/api/portal-api/models/author-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Prueba API
 * This Api will be responsible for overall data distribution
 *
 * The version of the OpenAPI document: v1
 * Contact: laihan758.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AuthorDto
 */
export interface AuthorDto {
    /**
     * 
     * @type {number}
     * @memberof AuthorDto
     */
    'id'?: number;
    /**
     * 
     * @type {number}
     * @memberof AuthorDto
     */
    'idBook'?: number;
    /**
     * 
     * @type {string}
     * @memberof AuthorDto
     */
    'firstName'?: string | null;
    /**
     * 
     * @type {string}
     * @memberof AuthorDto
     */
    'lastName'?: string | null;
}